+++ composer.json
@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": "^7.1",
+        "php": "^7.1|^8.0",
         "exsyst/swagger": "^0.4.1",
         "namshi/cuzzle": "^1|^2",
         "psr/http-message": "^1",
