+++ src/Factory/PactMessageFactory.php
@@ -24,35 +24,10 @@ namespace Bigfoot\PHPacto\Factory;
 use Bigfoot\PHPacto\Encoder\BodyEncoder;
 use Bigfoot\PHPacto\Encoder\HeadersEncoder;
 use Bigfoot\PHPacto\Matcher\Rules\EqualsRule;
-use Bigfoot\PHPacto\Matcher\Rules\Rule;
-use Bigfoot\PHPacto\Matcher\Rules\StringEqualsRule;
 use Psr\Http\Message\MessageInterface;
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
 
 abstract class PactMessageFactory
 {
-    protected static function getMethodRule(RequestInterface $request): Rule
-    {
-        return new EqualsRule(\strtoupper($request->getMethod()));
-    }
-
-    protected static function getUriRule(RequestInterface $request): Rule
-    {
-        $uri = $request->getUri()->getPath();
-
-        if ($request->getUri()->getQuery()) {
-            $uri .= $request->getUri()->getQuery();
-        }
-
-        return new StringEqualsRule($uri);
-    }
-
-    protected static function getStatusCodeRule(ResponseInterface $response): Rule
-    {
-        return new EqualsRule($response->getStatusCode());
-    }
-
     protected static function getHeadersRules(MessageInterface $response)
     {
         $decodedHeaders = HeadersEncoder::decode($response->getHeaders());
